Add package/{start,stop} endpoints
This commit adds endpoints for starting and stopping the release of a
package. As part of this, the post_packages method was altered to not
"start" the package, as it is not implicit that a package is started when
it is created. Thus thefull list of packages would be posted at
the start of a release, so that they can all be displayed on a
dashboard, then we indicate that it has started separately.

But this is all up to the user, there is nothing to stop someone
adding each package to a release as it is done.

I also changed renamend "success" column to "status", and made it an Enum
with specific values. Reason for the enum is enforcing a contract,
rather than any sort of optimisation (our data is not likely to be large
enough for this to matter).